requirements.txt
Flask>=3.0,<4
Werkzeug>=3.0,<4
gunicorn==22,<24
python-dotenv==0.21.0
requests==2.28.1
psycopg[binary]==3.2.3
psycopg-pool==3.2.6
